requirements.txt
Django==1.11.1
lxml==3.7.3
pkg-resources==0.0.0
psycopg2==2.7.1
python-magic==0.4.13
pytz==2017.2
requests==2.14.2
